Fixes to SSL/TLS analyzer
Analyzer can cope with zero length client and server
certificates.  It does still generate a weird though.
Extended cipherspec_size weirds are not thrown anymore,
they are incredibly overwhelming and should be handled
completely at the scripting in my opinion.

Integrated and expanded on patch Rmkml from ticket #209
that fixes problem with not parsing or expecting SSL
extensions.  SSL extensions still are not extracted
and passed to script land, but the analyzer doesn't
fail anymore.