Beginning of ssl.bro rewrite.  Far from working.
We may be fixing up the binpac ssl analyzer and getting
rid of the C++ one.  As I look closer and closer and at
C++ analyzer, I'm less impressed with the events it
outputs due to it maintaining state differently in the
core than most other analyzers.  Subsequently, the
events that it generates are also significantly
different from those of other analyzers.