A proof-of-concept opaque type.
This is just for playing around for the moment. It tries to address
the situation that a bif creates a C-level data structure that is then
to be manipulated with further bifs (but not with other script-level
operators). There are two problems with that:

    - We don't have a type at the script-level to represent the
    created data structure.

    - We can't just pass C-level pointers around (say, via "any")
    because then garbage collection wouldn't clean up after us.

This patch allows to wrap such a pointer with a record:

    type Foo = record {
        opaque: any;
        };

The bifs simply pass around instances of this record, while internally
storing a C pointer to the data structure instance in its only field.

Internally, "opaque" is an instance of a new Val-derived class
OpaqueVal. It gets cleaned up via ref-counting as usual and deletes
the C data structure automatically when done.