Refactoring for the win!  Pulled a lot of stuff out of LogMgr.h and moved it into other files.  Put all non-LogMgr stuff into a "bro" namespace; made a few changes to support that.
Added a static utility function that looks up a LogWriter by
type and launches a thread, so LogMgr's CreateWriter function is (much!)
shorter.

Updated LogWriterDS to use the new LogWriter (new constructor /
Initialize format).

Cleaned up header files a little bit, but still a lot of that left to
do.