Initial commit of log tools.  Point the tool at a directory and / or file, and the rest should be pretty automagic; it knows how to read gzip, bz2, standard log files, and DataSeries files.  Note that DataSeries files are expanded into a directory in /tmp; iterating through those without using a temporary file would require a python wrapper for DataSeries, since subprocess.PIPE is prone to race conditions when given large quantities of input / output.
The manager groups log files by the path bro used when they were
written.  Thus, manager["conn"] would provide a list of all the entires
for all the log entries corresponding to the bro 'conn' path found
anywhere in the directory tree.

Note that the ASCII log header has been updated / expanded somewhat.

Lots of stuff left to do, but it's a start.