Make CompositeHash key computation and value recovery deterministic.
Instead of being based on the memory address of the function Val, it's
now based on the function's identifier which can be used to recover the
function Val by looking it up in the global scope.

For anonymous functions, the temporary identifier is now unique per
function body definition and also is inserted into the global scope.

The previous non-determinism was mainly observed in the ordering of
the notice policy's elements as seen in the notice_policy.log.  The
new unit test establishes a baseline for that log which should no
longer differ between bro runs.