Almost done with event.bif.
Added comments to almost all events. The only ones I'm leaving out are
Gnutella and BitTorrent, don't know enough about those protocols to
document the events. If anybody does, please chime in.

A round of proofreading would be good too, there are probably still a
few typos in the texts.