Provide a new function split_esc.
This function extends the split* family with a new one, which is useful when
the split epxression may be escaped. For example, it is currently not possible
to split a string with ',' as delimiter where ',' itself may be escaped.
Specifically, the string

    "foo, bar\,baz, qux"

would yield 4 elements: "foo" and "bar\" and "baz" and "qux". With
split_escaped, one can specify a new escape character (here: \) to obtain a
result with just three elements: "foo" and "bar\,baz" and "qux".