Check SNI against CN subject field.
This check compares the SNI value (c$ssl$server_name) against the CN value in
the subject field of a certificate. If these do not match, it may mean that a
user sees a certificate error in the browser.

There is a caveat though. We still need to check the certificate's Server
Alternate Name (SAN) field to avoid false positives. The (optional) SAN fields
contains a list of domain names that the CN value should also be checked
against. However, Bro's TLS analyzer currently does not lift this field to the
scripting layer. Since Google uses the SAN field heavily, we need to extend the
TLS analyzer in order for this script to be deployable in practice.