Fix summary lines for BIF documentation
The summary lines (the first sentence in the description) for some BIFs
contained a period before the end of the sentence, so only part of the
sentence would appear in the "summary" section of the HTML document (fixed
by rewording the sentence).

Some summary lines were too long (fixed by splitting the sentence with first
sentence being more concise).

Also corrected the description of "fmt" and "floor" BIFs.